==> plasma_concentrations/src/03_format_extracted_unbound_datasets.ipynb <==
"""Clean up various characters etc from the extracted datasets of Fu/PPB from literature"""

#### RDKIT env ###

import pandas as pd
import re
from rdkit import Chem

basedir = '/scratch/ias41/ae_code/plasma_concentrations'

pd.set_option('display.max_rows', 1000)

### Dataset 1 - PMID15637086

dataset1 = pd.read_excel(basedir + '/data/PMID15637086_ETPC_from_HTML.xlsx')
dataset1.columns = [i.strip('\xa0') for i in dataset1.columns]

dataset1 = dataset1.applymap(lambda x: x.replace('\xa0','').replace('\u2002', '').replace('\u20021',''))

def remove_numbers(x):
    pattern = re.compile('[0-9]+\.(.*)')
    drug_name = re.match(pattern, x).group(1)
    return drug_name

dataset1['Drug name'] = dataset1['Drug name'].apply(lambda x: remove_numbers(x))

dataset1['ETCPun-bound (nM)'] = dataset1['ETCPun-bound (nM)'].apply(lambda x: x.replace(' ', ''))

# drop drug combination
dataset1 = dataset1.drop(42)

dataset1['Drug name'] = dataset1['Drug name'].apply(lambda x: x.replace(' i.v.', ''))

dataset1_selection = dataset1[['Drug name', 'ETCPun-bound (nM)']]
dataset1_selection.columns = ['Original drug name', 'ETCP unbound (nM)']
dataset1_selection['PMID'] = 'PMID15637086'

# Save to file
dataset1_selection.to_csv(basedir + '/results/interim/PMID15637086_data.txt', sep='\t', index=None)

### Dataset 2 - PMID22210121

dataset2 = pd.read_excel(basedir + '/data/PMID22210121_PPB_from_HTML.xlsx')

dataset2['fb(%)'] = dataset2['fb(%)'].apply(lambda x: str(x).replace('<', '').replace('>', ''))

def remove_approx(x):
    x = x.replace('\xa0', ' ').replace('\xa06', ' ')
    pattern = re.compile('([0-9\.]+) ± [0-9/.]+')
    match = re.match(pattern, x)
    if match:
        return match.group(1)
    else:
        return x

dataset2['fb(%)'] = dataset2['fb(%)'].apply(lambda x: remove_approx(x))

def find_upper_bound(x):
    match = re.match('([0-9]+)–([0-9]+)', str(x))
    if match:
        return float(match.group(1)), float(match.group(2))
    else:
        return float(x), None

dataset2['fb upper'] = dataset2['fb(%)'].apply(lambda x: find_upper_bound(x)[1])

dataset2['fb lower'] = dataset2['fb(%)'].apply(lambda x: find_upper_bound(x)[0])

dataset2_selection = dataset2[['Drugs', 'fb lower', 'fb upper']]
dataset2_selection['PMID'] = 'PMID22210121'
dataset2_selection.columns = ['Drug name', 'PPB lower', 'PPB upper', 'PMID']

def separate_synonyms(x):
    pattern = '(.+)\((.+)\)'
    match = re.match(pattern, x)
    if match:
        return match.group(1).strip(' '), match.group(2)
    else:
        return x, None

dataset2_selection['Original drug name'] = dataset2_selection['Drug name'].apply(lambda x: separate_synonyms(x)[0])
dataset2_selection['Original synonyms'] = dataset2_selection['Drug name'].apply(lambda x: separate_synonyms(x)[1])

# Save to file
dataset2_selection[['Original drug name', 'Original synonyms', 'PPB lower', 'PPB upper', 'PMID']].to_csv(basedir + '/results/interim/PMID22210121_data.txt', sep='\t', index=None)

### Dataset 3 - PMID30115648

dataset3 = pd.read_excel(basedir + '/data/PMID30115648_Fu.xlsx')

dataset3_selection = dataset3.loc[~dataset3['fraction unbound \nin plasma (fu)'].isnull(),:][['Name', 'SMILES', 'fraction unbound \nin plasma (fu)']]

dataset3_selection['Original drug name'] = dataset3_selection['Name'].apply(lambda x: separate_synonyms(x)[0])
dataset3_selection['Original synonyms'] = dataset3_selection['Name'].apply(lambda x: separate_synonyms(x)[1])

def smiles_to_inchikey(x):
    try:
        mol = Chem.MolFromSmiles(x)
        if mol:
            inchikey = Chem.InchiToInchiKey(Chem.inchi.MolToInchi(mol))
            return inchikey
    except NameError:
        return None

dataset3_selection['inchi_key'] = dataset3_selection['SMILES'].apply(lambda x: smiles_to_inchikey(x))

dataset3_selection.columns = ['Name', 'SMILES', 'Fu',
       'Original drug name', 'Original synonyms', 'inchi_key']

# save to file
dataset3_selection['PMID'] = 'PMID30115648'
dataset3_selection[['Original drug name', 'Original synonyms', 'Fu', 'PMID', 'inchi_key', 'SMILES']].to_csv(basedir + '/results/interim/PMID30115648_data.txt', sep='\t', index=None)

### Dataset 4 - Redfern et al.

dataset4 = pd.read_excel(basedir + '/data/PMID12667944_PPB.xlsx')

dataset4.drop(labels = ['ETPC ng/ml lower bound','ETPC ng/ml higher bound'], axis=1, inplace=True)

dataset4.columns = ['Original drug name', 'PPB', 'PPB upper', 'ETCP unbound (nM)',
       'ETCP unbound (nM) upper']

dataset4['PMID'] = 'PMID12667944'
dataset4.to_csv(basedir + '/results/interim/PMID12667944_data.txt', sep='\t', index=False)
