==> plasma_concentrations/src/04_mapping_unbound_data_to_approved.ipynb <==
"""Mapping compounds from extracted unbound concentrations, Fu, and Plasma protein binding to ChEMBL approved drugs"""

import pandas as pd
import numpy as np

basedir = '/scratch/ias41/ae_code/plasma_concentrations'

pd.set_option('display.max_colwidth',300)
pd.set_option('display.max_rows',300)

dataset1 = pd.read_csv(basedir + '/results/interim/PMID22210121_data.txt', sep='\t')
dataset2 = pd.read_csv(basedir + '/results/interim/PMID15637086_data.txt', sep='\t')
dataset3 = pd.read_csv(basedir + '/results/interim/PMID30115648_data.txt', sep='\t')
dataset4 = pd.read_csv(basedir + '/results/interim/PMID12667944_data.txt', sep='\t')
concatenated = pd.concat([dataset1, dataset2, dataset3, dataset4], ignore_index=True, sort=False)

concatenated.head()

drug_df = concatenated[['Original drug name', 'Original synonyms', 'inchi_key','PMID']].drop_duplicates()
drug_df['Original drug name'] = drug_df['Original drug name'].apply(lambda x: x.upper().strip(' '))
drug_df['Original synonyms'] = drug_df['Original synonyms'].apply(lambda x: str(x).upper().strip(' '))

drug_df.head()

approved = pd.read_csv(basedir + '/data/chembl_approved_drugs.txt', sep='\t')
approved['synonyms'] = approved['synonyms'].apply(lambda x: str(x).upper())

approved.head()

# Pref_name in approved are not unique - some compounds with different chembl_id have same name
len(approved[['chembl_id', 'pref_name']].drop_duplicates()), len(approved[['pref_name']].drop_duplicates())

### Round 1 - mapping on inchi key

inchi_key_mappings = pd.merge(drug_df, approved[['pref_name', 'chembl_id', 'molregno', 'standard_inchi_key', 'canonical_smiles']].drop_duplicates(), left_on = 'inchi_key', right_on='standard_inchi_key')
inchi_key_mappings = inchi_key_mappings.loc[~inchi_key_mappings['inchi_key'].isnull(),:].drop_duplicates()

inchi_key_mappings

### Round 2 - mapping on pref_name

pref_name_mappings = pd.merge(drug_df.loc[~drug_df['inchi_key'].isin(list(inchi_key_mappings['inchi_key'])),:], approved[['pref_name', 'chembl_id','molregno', 'standard_inchi_key', 'canonical_smiles']].drop_duplicates(), left_on = 'Original drug name', right_on='pref_name').drop_duplicates()
pref_name_mappings

all_mappings = pd.concat([inchi_key_mappings, pref_name_mappings],sort=False,ignore_index=True).drop_duplicates()

all_mappings.sort_values(by='pref_name')

### Round 3 - mapping of source synonym to chembl pref_name

unmapped = drug_df.loc[(~drug_df['Original drug name'].isin(list(all_mappings['Original drug name'])))&(~drug_df['inchi_key'].isin(list(all_mappings['standard_inchi_key']))),:]

len(unmapped)

synonym_pref_name_mappings_initial = pd.merge(drug_df, approved[['pref_name', 'chembl_id','molregno', 'standard_inchi_key','canonical_smiles']].drop_duplicates(), left_on='Original synonyms', right_on='pref_name').drop_duplicates().sort_values(by='pref_name')

synonym_pref_name_mappings_initial

# Round 4

# Original drug name to ChEMBL synonyms
substance_synonyms_mappings = pd.merge(unmapped, approved, left_on='Original drug name', right_on='synonyms')[['pref_name','Original drug name','syn_type','chembl_id','molregno', 'standard_inchi_key','canonical_smiles']].drop_duplicates()

# Inspect manually and decide which ones to drop
substance_synonyms_mappings.sort_values(by='Original drug name')

# accept all but tenofovir and candesartan - these are phase 3
to_drop = ['CHEMBL1014', 'CHEMBL1486']
substance_synonyms_mappings_to_add = substance_synonyms_mappings.loc[~substance_synonyms_mappings['chembl_id'].isin(to_drop),all_mappings.columns].drop_duplicates()

substance_synonyms_mappings_to_add

len(all_mappings)

all_mappings = pd.concat([all_mappings, substance_synonyms_mappings_to_add], sort=False,ignore_index=True).drop_duplicates()

len(all_mappings)

all_mappings.head()

# Round 5

unmapped_2 = drug_df.loc[(~drug_df['Original drug name'].isin(list(all_mappings['Original drug name'])))&(~drug_df['inchi_key'].isin(list(all_mappings['standard_inchi_key']))),:]

len(unmapped_2)

unmapped_2.head()

# Substance to compound name
# Original drug name to ChEMBL synonyms
round5_mappings = pd.merge(unmapped_2, approved, left_on='Original drug name', right_on='compound_name')[['pref_name','Original drug name','chembl_id','molregno',]].drop_duplicates()

round5_mappings

# Round 6
Synonyms to synonyms

# Synonyms to synonyms
syn_to_syn_mappings = pd.merge(unmapped_2.loc[~unmapped_2['Original synonyms'].isnull(),:], approved, left_on='Original synonyms',right_on='synonyms')[['Original drug name','pref_name','Original synonyms','syn_type','chembl_id','molregno', 'standard_inchi_key','canonical_smiles']].drop_duplicates()
syn_to_syn_mappings = syn_to_syn_mappings.loc[syn_to_syn_mappings['Original synonyms']!='NAN',:]

syn_to_syn_mappings

len(all_mappings)

all_mappings.loc[all_mappings['Original drug name']=='DOXYCYCLINE']

# Checking that all 'substance' only have one mapping to chembl_id
for name in list(set(all_mappings['Original drug name'])):
    subset_df = all_mappings.loc[all_mappings['Original drug name']==name,'molregno'].drop_duplicates()
    if len(subset_df)>1:
        print(name)

# Drop non-parents
to_be_dropped = ['CHEMBL1200788', 'CHEMBL1200544', 'CHEMBL1646', 'CHEMBL3184906', 'CHEMBL3989887', 'CHEMBL2103745', 'CHEMBL1201018', 'CHEMBL3989676', 'CHEMBL1200610', 'CHEMBL1200338', 'CHEMBL1200699']

all_mappings = all_mappings.loc[~all_mappings['chembl_id'].isin(to_be_dropped),:]

len(all_mappings)

# Checking that all 'substance' only have one mapping to chembl_id
for name in list(set(all_mappings['Original drug name'])):
    subset_df = all_mappings.loc[all_mappings['Original drug name']==name,'molregno'].drop_duplicates()
    if len(subset_df)>1:
        print(name)

# See currently unmapped

unmapped_3 = drug_df.loc[(~drug_df['Original drug name'].isin(list(all_mappings['Original drug name'])))&(~drug_df['inchi_key'].isin(list(all_mappings['standard_inchi_key']))),:]

unmapped_3.head()

unmapped_3['short substance'] = unmapped_3['Original drug name'].apply(lambda x: x.rstrip('E'))

unmapped_3.head()

## Attach mappings after considering trailing 'e'

# Noticed some names had an 'e' on the end that prevented synonym mapping
my_columns = [i for i in all_mappings.columns]
short_names = pd.merge(unmapped_3, approved, left_on='short substance', right_on='pref_name')[my_columns].drop_duplicates()
short_names

len(set(drug_df['Original drug name'])), len(set(all_mappings['Original drug name'])), len(set(all_mappings['molregno']))

all_mappings.head()

# Need to match back to original datasets

ds1_mappings = all_mappings.loc[all_mappings['PMID']=='PMID22210121',:][['Original drug name', 'chembl_id', 'molregno']]
dataset1['Original drug name'] = dataset1['Original drug name'].apply(lambda x: x.upper().strip(' '))
ds1_result = ds1_mappings.merge(dataset1, on='Original drug name', how='inner')
ds1_result.to_csv(basedir + '/results/interim/PMID22210121_mapped_data.txt', sep='\t', index=False)

ds2_mappings = all_mappings.loc[all_mappings['PMID']=='PMID15637086',:][['Original drug name', 'chembl_id', 'molregno']]
dataset2['Original drug name'] = dataset2['Original drug name'].apply(lambda x: x.upper().strip(' '))
ds2_result = ds2_mappings.merge(dataset2, on='Original drug name', how='inner')
ds2_result.to_csv(basedir + '/results/interim/PMID15637086_mapped_data.txt', sep='\t', index=False)

ds3_mappings = all_mappings.loc[all_mappings['PMID']=='PMID30115648',:][['Original drug name', 'chembl_id', 'molregno']]
dataset3['Original drug name'] = dataset3['Original drug name'].apply(lambda x: x.upper().strip(' '))
ds3_result = ds3_mappings.merge(dataset3, on='Original drug name', how='inner')
ds3_result.to_csv(basedir + '/results/interim/PMID30115648_mapped_data.txt', sep='\t', index=False)

ds4_mappings = all_mappings.loc[all_mappings['PMID']=='PMID12667944',:][['Original drug name', 'chembl_id', 'molregno']]
dataset4['Original drug name'] = dataset4['Original drug name'].apply(lambda x: x.upper().strip(' '))
ds4_result = ds4_mappings.merge(dataset4, on='Original drug name', how='inner')
ds4_result.to_csv(basedir + '/results/interim/PMID12667944_mapped_data.txt', sep='\t', index=False)
