==> plasma_concentrations/src/06_integrate_plasma_conc_and_unbound_data_medians.ipynb <==
"""Take mean of lower-upper Fu/PPB per publication, take medians of Fus from different publications, calculate Fu and multiply by total plasma conc. to get unbound. Add the few extracted unbound concentrations back and take median per compound. Add back a few total concentrations without unbound info."""

import pandas as pd
import numpy as np

1. if unbound concentration available, keep apart to combine later
2. for other data, take median of various fu/ppb figures
3. merge that data with plasma concentrations and use to calculate 'calculated unbound concentration'
4. combined with etcp data
4. convert to -log(molar unbound) 


basedir = '/scratch/ias41/ae_code/plasma_concentrations'

combined_data = pd.read_csv(basedir + '/results/interim/combined_fu_ppb_etcp.txt', sep='\t')
plasma_conc = pd.read_csv(basedir + '/results/interim/molregno2median_total_plasma_conc.txt', sep='\t')
approved_drugs = pd.read_csv(basedir + '/data/chembl_approved_drugs.txt', sep='\t')

molregno2pref_name = dict()
for row in approved_drugs[['molregno','pref_name']].drop_duplicates().iterrows():
    molregno2pref_name[row[1]['molregno']] = row[1]['pref_name']

molregno2pref_name[97]

plasma_conc.head()

plasma_conc.columns = ['molregno', 'pref_name','chembl_id','median Molar total plasma concentration', 'median pMolar total plasma concentration']

etcp_data = combined_data.loc[combined_data['type']=='ETCP unbound (nM)',:]
other_data = combined_data.loc[combined_data['type']!='ETCP unbound (nM)',:]

#### For ETCP data, convert to concentrations to Molar value. If there are two values, take mean

def value_nM_to_M(row):
    # Take higher range value
    if pd.isna(row['upper_value']):
        return (row['value'] * (10**-9))
    if pd.isna(row['upper_value']) == False:
        return (np.mean([row['value'], row['upper_value']]) * (10**-9))

etcp_data['unbound_molar_value'] = etcp_data.apply(value_nM_to_M, axis=1)

#### Take mean of available upper and lower PPB measurements per compound from same publication

# Invert PPB
def calculate_fu(x):
    if x['type'] == 'PPB':
        if pd.isna(x['upper_value']):
            Fu = (100 - x['value'])/100
        elif pd.isna(x['upper_value']) == False:
            # in case a range is reported - take average of value and upper_value
            ppb_average = (x['value']+x['upper_value'])/2
            fu = Fu = (100 - ppb_average)/100
        return Fu
    elif x['type'] == 'Fu':
        return x['value']

other_data['Fu'] = other_data.apply(calculate_fu, axis=1)

# Take median of Fu data from different sources
#Fus = other_data.groupby(by='molregno')['Fu'].median().reset_index()

other_data.sort_values(by='molregno')

# Merge with plasma concentrations
Fus_merged = other_data[['molregno','Fu']].merge(plasma_conc, on='molregno')

Fus_merged.loc[Fus_merged['chembl_id'].isnull()]

# Calculate unbound concentration
Fus_merged['Molar unbound plasma concentration'] = Fus_merged['Fu'] * Fus_merged['Molar total plasma concentration']

etcp_data.columns

etcp_data.loc[etcp_data['chembl_id'].isnull()]

# combine with etcp data
ds1 = etcp_data[['molregno', 'chembl_id','unbound_molar_value']]
ds1.columns = ['molregno','chembl_id','Molar unbound plasma concentration']

ds1['pref_name'] = ds1['molregno'].apply(lambda x: molregno2pref_name[x])

Fus_merged.columns

ds2 = Fus_merged[['pref_name','molregno','chembl_id','Molar total plasma concentration','Molar unbound plasma concentration']]
combined_unbound = pd.concat([ds1,ds2], ignore_index=True, sort=False)

combined_unbound.head()

combined_unbound.tail()

# Now, after adding unbound datapoints, calculate median
# Some of the total columns have NA values because some values were reported as unbound directly
# The default of 'median' is to skip NAs so they are not included in calculation
medians_unbound = combined_unbound.groupby('molregno').agg({'Molar unbound plasma concentration': 'median', 'chembl_id': 'first', 'pref_name': 'first', 'Molar total plasma concentration': 'median'})

medians_unbound.rename(columns = {'Molar unbound plasma concentration': 'median Molar unbound plasma concentration', 'Molar total plasma concentration': 'median Molar total plasma concentration'}, inplace=True)

medians_unbound.reset_index(drop=False, inplace=True)

medians_unbound.columns

# Add back in the rows without unbound from total dataframe (which were already medians)
# Rows from total plasma without unbound data
total_only_rows = plasma_conc.loc[~plasma_conc['molregno'].isin(medians_unbound['molregno'])].sort_values(by='pref_name')
total_only_rows.head()

medians_total_unbound = pd.concat([medians_unbound, total_only_rows], sort=False)

medians_total_unbound.loc[medians_total_unbound['median Molar unbound plasma concentration'].isnull()]

len(medians_total_unbound['molregno'].drop_duplicates())

# calulate -log(molar unbound concentration)
medians_total_unbound['median pMolar total plasma concentration'] = medians_total_unbound['median Molar total plasma concentration'].apply(lambda x: -np.log10(x))
medians_total_unbound['median pMolar unbound plasma concentration'] = medians_total_unbound['median Molar unbound plasma concentration'].apply(lambda x: -np.log10(x))

medians_total_unbound.head()

# Save to CSV
medians_total_unbound[['molregno', 'chembl_id', 'pref_name', 'median Molar total plasma concentration', 'median pMolar total plasma concentration', 'median Molar unbound plasma concentration', 'median pMolar unbound plasma concentration']].to_csv(basedir + '/results/molregno2median_plasma_total_unbound.txt', sep='\t', index=False)
